fix: rename from 'mspr' to 'mspr_api'
