+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-spring.application.name=mspr
+spring.application.name=mspr_api
