+++ src/test/java/fr/mspr_api/MsprAPIApplicationTests.java
@@ -1,10 +1,10 @@
-package fr.mspr;
+package fr.mspr_api;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class MsprApplicationTests {
+class MsprAPIApplicationTests {
 
 	@Test
 	void contextLoads() {
